+++ src/pages/Settings.jsx
@@ -4,7 +4,6 @@ import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
 import { BsChevronDown } from 'react-icons/bs';
 
-
 const Settings = (props) => {
     const [displayDropdown, setDisplayDropdown] = useState(false);
     const navigate = useNavigate();